pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "ml_inference_service"
version = "0.1.0"
description = "ML inference API with billing and promocodes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "psycopg2-binary>=2.9.10",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.17",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "joblib>=1.4.2",
    "scikit-learn>=1.5.0",
    "numpy>=2.0.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "httpx>=0.28.0",
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]
streamlit = [
    "streamlit>=1.40.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["ml_inference_service*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["ml_inference_service"]
omit = ["*/tests/*", "*/alembic/*"]

[tool.coverage.report]
fail_under = 70
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]
